requirements.txt
datasets
tkinter
pandas
sklearnmatplotlib
nltk
